refactor(server): extract pure helpers into src/utils
