+++ packages/emitsignal-server/src/http/subscriptions/subscribe.ts
@@ -2,9 +2,9 @@ import Elysia, { t } from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { prisma } from '#/lib/prisma';
-import { serializeSubscriptionSettings } from '#/lib/subscription-settings';
 import { getOrCreateTopic, TOPIC_NAME_MAX_LENGTH, TopicNameError } from '#/lib/topic';
 import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { serializeSubscriptionSettings } from '#/utils/subscription-settings';
 
 export const subscribe = new Elysia({ prefix: '/subscriptions' }).post(
     '/',
