+++ packages/emitsignal-server/src/http/topic/metrics.ts
@@ -2,10 +2,10 @@ import Elysia from 'elysia';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { authAwareBeforeHandle } from '#/http/plugins/rate-limit-plugin';
-import { duration } from '#/lib/duration';
 import { prisma } from '#/lib/prisma';
 import { readAnonLimiter, readAuthLimiter } from '#/lib/rate-limit';
 import { resolveTopicCapabilities } from '#/lib/topic-access';
+import { duration } from '#/utils/duration';
 
 export const topicMetrics = new Elysia().get(
     '/topics/:name/metrics',
