+++ packages/emitsignal-server/src/http/webhooks/list.ts
@@ -2,8 +2,8 @@ import Elysia from 'elysia';
 
 import { Prisma } from '#/generated/prisma/client';
 import { resolveUserId } from '#/http/auth/plugin';
-import { duration } from '#/lib/duration';
 import { prisma } from '#/lib/prisma';
+import { duration } from '#/utils/duration';
 
 export const listWebhooks = new Elysia().get('/webhooks', async ({ headers, status }) => {
     const userId = await resolveUserId({ headers });
