+++ packages/emitsignal-server/src/http/webhooks/receive.ts
@@ -1,11 +1,10 @@
 import { parseTemplate, renderTemplate } from '@emitsignal/shared/webhook-template';
 import Elysia, { t } from 'elysia';
 
-import type { Action } from '#/lib/actions';
+import type { Action } from '#/utils/actions';
 
 import { resolveUserId } from '#/http/auth/plugin';
 import { consumeLimit } from '#/http/plugins/rate-limit-plugin';
-import { validateActions } from '#/lib/actions';
 import { getUserPlan } from '#/lib/billing/get-user-plan';
 import { messageExpiresAt, messageRetentionDays } from '#/lib/billing/retention';
 import { bus } from '#/lib/event-bus';
@@ -14,6 +13,7 @@ import { pushQueue } from '#/lib/queue';
 import { webhookReceiveLimiter } from '#/lib/rate-limit';
 import { getOrCreateTopic, serializeMessage } from '#/lib/topic';
 import { canPublishToTopicName } from '#/lib/topic-access';
+import { validateActions } from '#/utils/actions';
 
 // Public endpoint: cap the inbound payload so a single caller can't store huge
 // WebhookDelivery rows or exhaust memory. 64 KB is generous for webhook JSON.
