+++ packages/emitsignal-server/src/lib/__tests__/media-refs.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
 
-import { normalizeMediaInput, validateMessageMedia } from '#/lib/media-refs';
+import { normalizeMediaInput, validateMessageMedia } from '#/utils/media-refs';
 
 describe('normalizeMediaInput', () => {
     it('returns empty for null/undefined', () => {
