+++ packages/emitsignal-server/src/lib/queue/index.ts
@@ -1,9 +1,9 @@
 export { redisConnection } from './connection';
-export { emailQueue } from './email/email-queue';
-export { createEmailWorker } from './email/email-worker';
-export { purgeQueue, scheduleRetentionSweep } from './purge/purge-queue';
-export { createPurgeWorker } from './purge/purge-worker';
-export { pushQueue } from './push/push-queue';
-export { createPushWorker } from './push/push-worker';
-export { scheduleQueue } from './schedule/schedule-queue';
-export { createScheduleWorker } from './schedule/schedule-worker';
+export { emailQueue } from '#/lib/queue/email/email-queue';
+export { createEmailWorker } from '#/lib/queue/email/email-worker';
+export { purgeQueue, scheduleRetentionSweep } from '#/lib/queue/purge/purge-queue';
+export { createPurgeWorker } from '#/lib/queue/purge/purge-worker';
+export { pushQueue } from '#/lib/queue/push/push-queue';
+export { createPushWorker } from '#/lib/queue/push/push-worker';
+export { scheduleQueue } from '#/lib/queue/schedule/schedule-queue';
+export { createScheduleWorker } from '#/lib/queue/schedule/schedule-worker';
