+++ packages/emitsignal-server/src/lib/queue/schedule/schedule-queue.ts
@@ -1,7 +1,7 @@
 import { ConnectionOptions, Queue } from 'bullmq';
 
-import { duration } from '#/lib/duration';
 import { redisConnection } from '#/lib/queue/connection';
+import { duration } from '#/utils/duration';
 
 export interface ScheduleJob {
     messageId: string;
