+++ packages/emitsignal-server/src/lib/storage/s3-provider.ts
@@ -1,4 +1,4 @@
-import { duration } from '#/lib/duration';
+import { duration } from '#/utils/duration';
 
 import type { FileMetadata, FileStorage, FileUploadInput, StorageBucket } from './provider';
