+++ packages/emitsignal-server/src/utils/duration.ts
