+++ packages/emitsignal-server/src/utils/subscription-settings.ts
